requirements.txt
# requirements.txt records the full set of dependencies for development
accelerate
codetiming
datasets
dill
hydra-core
liger-kernel
numpy
pandas
datasets
peft
pyarrow>=15.0.0
pybind11
pylatexenc
pylint==3.3.6
ray[default]
tensordict<=0.6.2
torchdata
transformers==4.51.2
vllm==0.8.5
swanlab
jsonlines